Move parse_type_ann
