+++ src/parser/src/lib.rs
@@ -10,6 +10,7 @@ pub mod ident;
 pub mod if_else;
 pub mod let_stmt;
 pub mod parse_loop;
+pub mod parse_type_ann;
 pub mod seq;
 pub mod structs;
 
@@ -150,39 +151,6 @@ impl<'inp> Parser<'inp> {
             ))
         }
     }
-
-    /// Parse and return type annotation. Expect lexer.peek() to be at Colon before call
-    // Should only consume tokens belonging to the annotation, starting peek at first token and ending
-    // peek at the last token of the annotation
-    fn parse_type_annotation(&mut self) -> Result<Type, ParseError> {
-        // self.consume_token_type(Token::Colon, "Expected a colon")?;
-        // expect_token_body!(self.lexer.peek(), Ident, "identifier")?;
-        Parser::expect_token_for_type_ann(self.lexer.peek())?;
-
-        // if ident, get the string and try to convert type. else, handle specially
-        let peek = self
-            .lexer
-            .peek()
-            .unwrap()
-            .to_owned()
-            .expect("Lexer should not fail"); // would have erred earlier
-
-        let type_ann = match peek {
-            Token::Ident(id) => Type::from_string(&id),
-            Token::OpenParen => {
-                self.advance();
-                if let Some(Ok(Token::CloseParen)) = self.lexer.peek() {
-                    Ok(Type::Unit)
-                } else {
-                    Err(ParseError::new("Expected '()' for unit type annotation"))
-                }
-            }
-            _ => unreachable!(),
-        }?;
-
-        Ok(type_ann)
-    }
-
     /* Precedence */
 
     // Return (left bp, right bp)
@@ -385,39 +353,6 @@ mod tests {
     }
 
     #[test]
-    fn test_parse_type_annotations() {
-        test_parse("let x : int = 2;", "let x : int = 2;");
-        test_parse("let x : bool = true;", "let x : bool = true;");
-        test_parse("let x : float = true;", "let x : float = true;");
-        test_parse("let x : () = true;", "let x : () = true;");
-    }
-
-    #[test]
-    fn test_parse_type_annotations_errs() {
-        // test_parse("let x : int = 2;", "");
-        test_parse_err(
-            "let x : let ",
-            "Expected identifier or '(' for type annotation, got 'let'",
-            true,
-        );
-        test_parse_err(
-            "let x : 2 ",
-            "Expected identifier or '(' for type annotation, got '2'",
-            true,
-        );
-        test_parse_err(
-            "let x : ",
-            "Expected identifier or '(' for type annotation, got end of input",
-            true,
-        );
-        test_parse_err(
-            "let x : (2 ",
-            "Expected '()' for unit type annotation",
-            true,
-        );
-    }
-
-    #[test]
     fn test_parse_concurrency() {
         let t = r"
         let t = spawn func();
@@ -487,5 +422,12 @@ mod tests {
 
         let t = r#"let t = "hello world"; println(t);"#;
         test_parse(t, "let t = hello world;println(t);");
+    }
+
+    #[test]
+    fn test_parse_type_annotations_more() {
+        let t = r"
+        let g : fn(int) -> int = f;
+        ";
     }
 }
